+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Upgraded BepInEx dependency.
 - Implemented an emissive visual warning when the nuke has only one hit remaining.
+- The nuke now completely destroys itself upon explosion to prevent the explosion from getting spammed every update, especially on lower strengths.
 - Changed config entry `HitSensitivity` to use an enum value to make it more readable and intuitive, particularly with REPOConfig.
 - README changes:
   - Updated the Configuration section to my new table style;